requirements.txt
wheel
https://github.com/KurimuzonAkuma/pyrogram/archive/dev.zip
tgcrypto
pushbullet.py
uvloop
